backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and LangGraph
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
langchain-core==0.1.0

# Reddit API
asyncpraw==7.7.1

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Ollama integration
ollama==0.1.7
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2

# Text processing and NLP
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
tiktoken==0.5.2
emoji==2.8.0
beautifulsoup4==4.12.2

# Database
sqlite3  # Built-in with Python
sqlalchemy==2.0.23

# Async support
asyncio-throttle==1.0.2
aiofiles==23.2.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
